requirements.txt
absl-py==0.12.0
antialiased-cnns==0.3
anyio==2.2.0
argcomplete==1.12.3
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
atari-py==0.2.6
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
boto==2.49.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
crcmod==1.7
cryptography==3.4.7
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
dominate==2.6.0
EasyProcess==0.3
entrypoints==0.3
ez-setup==0.9
fasteners==0.16
Flask==2.0.1
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gcs-oauth2-boto-plugin==2.7
gin-config==0.4.0
google-apitools==0.5.31
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
google-reauth==0.1.1
grpcio==1.32.0
gsutil==4.61
gym==0.18.0
gym-minigrid==1.0.2
h5py==2.10.0
HTML4Vision==0.4.2
httplib2==0.19.1
idna==2.10
imageio==2.9.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.9.2
jupyter-server==1.6.3
jupyterlab==3.0.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
learn2learn==0.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.1
mistune==0.8.4
mock==2.0.0
monotonic==1.6
nbclassic==0.2.7
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.3.0
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.8
psutil==5.8.0
ptyprocess==0.7.0
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyglet==1.5.0
Pygments==2.8.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyu2f==0.1.5
PyVirtualDisplay==2.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
retry-decorator==1.1.1
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.7.1
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
stable_baselines3==1.1
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-hub==0.12.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.4.8
tomlkit==0.7.0
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
trash-cli==0.21.6.10.1
typing-extensions==3.7.4.3
Unidecode==1.2.0
urllib3==1.26.4
v==0.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yacs==0.1.8
